feat: update Next.js configuration for static export and environment variables
refactor: remove local data fetching and image downloading scripts

feat: implement BlogPageClient for client-side filtering and searching

fix: update blog post fetching to use Notion API directly

chore: clean up unused utility functions related to local data and image downloading

style: improve code readability and organization in blog components

fix: adjust LandingPage to handle errors and loading states more gracefully
