+++ package.json
@@ -7,7 +7,6 @@
   "type": "module",
   "scripts": {
     "dev": "next dev --turbopack",
-    "prebuild": "bun run scripts/build-data.ts",
     "build": "next build",
     "start": "next start",
     "test": "vitest",
@@ -41,7 +40,6 @@
   },
   "dependencies": {
     "@notionhq/client": "^4.0.0",
-    "@tanstack/react-query": "^5.81.5",
     "framer-motion": "^12.23.0",
     "katex": "^0.16.22",
     "lucide-react": "^0.525.0",
