+++ src/app/page.tsx
@@ -1,39 +1,39 @@
 import React from 'react';
 
 import { LandingTemplate } from '@/components/templates/LandingTemplate';
-import { getLocalPostMeta } from '@/utils/localData';
-import { extractCategories } from '@/utils/blog';
+import { getPostList } from '@/libs/notion';
 
 /**
  * 메인 페이지 (Landing Page)
- * SSG (Static Site Generation)를 사용하여 빌드 타임에 데이터를 가져옵니다
+ * 정적 생성으로 빌드 타임에 데이터 가져오기
  */
 export default async function LandingPage() {
   try {
-    // 빌드된 로컬 데이터에서 포스트 목록 가져오기 (SSG)
-    const posts = await getLocalPostMeta();
+    // 빌드 타임에 데이터 가져오기
+    const posts = await getPostList();
 
-    // 카테고리 목록 추출
-    const categories = extractCategories(posts);
+    // 카테고리 추출
+    const categories = ['전체', ...Array.from(new Set(posts.map(post => post.category.text)))];
 
     return (
       <LandingTemplate
         posts={posts}
         categories={categories}
+        isLoading={false}
         showMoreButton={true}
         moreButtonText="둘러보기"
-        emptyMessage="아직 포스트가 없습니다."
+        emptyMessage=""
       />
     );
   } catch (error) {
     console.error('Failed to load posts:', error);
-
-    // 에러 발생 시 빈 상태로 렌더링
     return (
       <LandingTemplate
         posts={[]}
         categories={['전체']}
-        showMoreButton={false}
+        isLoading={false}
+        showMoreButton={true}
+        moreButtonText="둘러보기"
         emptyMessage="포스트를 불러오는 중 오류가 발생했습니다."
       />
     );
