+++ src/components/ui/blog/PostCard/PostCard.tsx
@@ -6,7 +6,6 @@ import { Heading, Text, DateText } from '@/components/ui/typography';
 import { TagButton } from '@/components/ui/buttons/TagButton';
 import { LabelButton } from '@/components/ui/buttons/LabelButton';
 import { PostMeta } from '@/types/blog';
-import { formatPostDate } from '@/utils/blog';
 
 /**
  * PostCard 컴포넌트 Props
@@ -80,7 +79,7 @@ export const PostCard = ({
 
         {/* 날짜 */}
         <div className="mb-3 text-left">
-          <DateText>{formatPostDate(createdAt)}</DateText>
+          <DateText>{createdAt}</DateText>
         </div>
 
         {/* 설명 */}
