+++ src/components/ui/blog/PostHeader/PostHeader.tsx
@@ -5,7 +5,6 @@ import { LabelButton } from '@/components/ui/buttons/LabelButton';
 import { TagButton } from '@/components/ui/buttons/TagButton';
 import { Heading, DateText, Text } from '@/components/ui/typography';
 import { PostMeta } from '@/types/blog';
-import { formatPostDate } from '@/utils/blog';
 
 export interface PostHeaderProps extends Omit<PostMeta, 'href'> {
   /**
@@ -78,7 +77,7 @@ export const PostHeader = ({
 
       {/* 날짜 */}
       <div className="mb-6">
-        <DateText>{formatPostDate(createdAt)}</DateText>
+        <DateText>{createdAt}</DateText>
       </div>
 
       {/* 태그 목록 */}
